refactor: rename CurrentScenario to ThreadStaticStepCollection and add missing docs
