+++ src/Xbehave.2/Sdk/Step.cs
@@ -10,7 +10,10 @@ namespace Xbehave.Sdk
     using System.Threading.Tasks;
 
     /// <summary>
-    /// Provides the implementation to execute each step.
+    /// Provides the natural language associated with a step, the body of a step,
+    /// the teardowns to be invoked after the execution of the scenario in which the step participates,
+    /// the objects to be disposed after the execution of the scenario in which the step participates and
+    /// a reason for skipping this step.
     /// </summary>
     [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Step", Justification = "By design.")]
     public class Step
@@ -20,6 +23,11 @@ namespace Xbehave.Sdk
         private readonly List<IDisposable> disposables = new List<IDisposable>();
         private readonly List<Action> teardowns = new List<Action>();
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Step"/> class.
+        /// </summary>
+        /// <param name="text">The natural language associated with step.</param>
+        /// <param name="body">The body of the step.</param>
         public Step(string text, Action body)
             : this(text)
         {
@@ -30,6 +38,11 @@ namespace Xbehave.Sdk
             };
         }
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Step"/> class.
+        /// </summary>
+        /// <param name="text">The natural language associated with step.</param>
+        /// <param name="body">The body of the step.</param>
         public Step(string text, Func<Task> body)
             : this(text)
         {
@@ -41,41 +54,64 @@ namespace Xbehave.Sdk
             this.text = text;
         }
 
+        /// <summary>
+        /// Gets the natural language associated with step.
+        /// </summary>
         public virtual string Text
         {
             get { return this.text; }
         }
 
+        /// <summary>
+        /// Gets the body of the step.
+        /// </summary>
         public virtual Func<object> Body
         {
             get { return this.body; }
         }
 
-        public IReadOnlyList<IDisposable> Disposables
-        {
-            get { return this.disposables.ToArray(); }
-        }
-
+        /// <summary>
+        /// Gets the teardowns to be invoked after the execution of the scenario in which the step participates.
+        /// </summary>
         public IReadOnlyList<Action> Teardowns
         {
             get { return this.teardowns.ToArray(); }
         }
 
-        public string SkipReason { get; set; }
-
-        public void AddDisposable(IDisposable disposable)
+        /// <summary>
+        /// Gets the objects to be disposed after the execution of the scenario in which the step participates.
+        /// </summary>
+        public IReadOnlyList<IDisposable> Disposables
         {
-            if (disposable != null)
-            {
-                this.disposables.Add(disposable);
-            }
+            get { return this.disposables.ToArray(); }
         }
 
-        public void AddTeardown(Action teardown)
+        /// <summary>
+        /// Gets or sets the reason for skipping this step.
+        /// </summary>
+        public string SkipReason { get; set; }
+
+        /// <summary>
+        /// Adds a teardown to be invoked after the execution of the scenario in which the step participates.
+        /// </summary>
+        /// <param name="teardown">The body of the teardown.</param>
+        public void Add(Action teardown)
         {
             if (teardown != null)
             {
                 this.teardowns.Add(teardown);
+            }
+        }
+
+        /// <summary>
+        /// Adds an object to be disposed after the execution of the scenario in which the step participates.
+        /// </summary>
+        /// <param name="disposable">A disposable object.</param>
+        public void Add(IDisposable disposable)
+        {
+            if (disposable != null)
+            {
+                this.disposables.Add(disposable);
             }
         }
     }
