+++ src/Xbehave.2/Xbehave.2.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -72,7 +72,7 @@
     <Compile Include="Fluent\Step.cs" />
     <Compile Include="IStepContext.cs" />
     <Compile Include="ScenarioAttribute.cs" />
-    <Compile Include="Sdk\CurrentScenario.cs" />
+    <Compile Include="Sdk\ThreadStaticStepCollection.cs" />
     <Compile Include="Sdk\Guard.cs" />
     <Compile Include="Sdk\Step.cs" />
     <Compile Include="StringExtensions.Async.cs" />
